refacto: move business files in their own folder
