+++ src/features/home/components/modules/business/helpers/getBusinessUrl.ts
@@ -8,18 +8,3 @@ export const getBusinessUrl = (url: string, email?: string) => {
 
   return email ? url.replace(emailFieldRegEx, email) : null
 }
-
-export const showBusinessModule = (
-  targetNotConnectedUsersOnly: boolean | undefined,
-  connected: boolean
-): boolean => {
-  // Target both type of users
-  if (targetNotConnectedUsersOnly === undefined) return true
-
-  // Target only NON connected users
-  if (!connected && targetNotConnectedUsersOnly) return true
-
-  // Target only connected users
-  if (connected && !targetNotConnectedUsersOnly) return true
-  return false
-}
