+++ src/features/home/components/modules/business/helpers/showBusinessModule.ts
@@ -1,0 +1,15 @@
+export const showBusinessModule = (
+  targetNotConnectedUsersOnly: boolean | undefined,
+  connected: boolean
+): boolean => {
+  // Target both type of users
+  if (targetNotConnectedUsersOnly === undefined) return true
+
+  // Target only NON connected users
+  if (!connected && targetNotConnectedUsersOnly) return true
+
+  // Target only connected users
+  if (connected && !targetNotConnectedUsersOnly) return true
+
+  return false
+}
